lambda_requirements.txt
fastapi==0.104.1
mangum>=0.17.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
loguru==0.7.2
